+++ usart.h
@@ -15,7 +15,7 @@
 void vUSART_init();
 
 void vUSART_send(uint8_t *data, uint16_t len);
-void vUSART_set_receive_callback(void(*cb)(uint8_t*, uint16_t));
+void vUSART_set_receive_callback(void(*cb)(uint8_t*, uint8_t));
 
 #endif /* USART_H_ */
